requirements.txt
dspy-ai
google-generativeai
python-dotenv
requests

